Client-Side Validation & Static Reorganization
