+++ views/account.handlebars
@@ -5,14 +5,19 @@
 <main>
 
   <form id="accountHandling">
-    <label>
-      Login<br><br />
-      <label id='formLabel' for="loginUser">Username:</label><br>
-      <input class='formText' type="text" id="loginUser" name="loginUser"><br><br />
-      <label id='formLabel' for="loginPassword">Password:</label><br>
-      <input class='formText' type="password" id="loginPassword" name="loginPassword"><br><br />
-      <input class="submit" type="submit" value="Login">
-    </label>
+    Login<br><br>
+
+    <label class='formLabel' for="loginUser">
+      Username:<br>
+      <input class='formText' type="text" id="loginUser" name="loginUser">
+    </label><br><br>
+
+    <label class='formLabel' for="loginPassword">
+      Password:<br>
+      <input class='formText' type="text" id="loginPassword" name="loginPassword">
+    </label><br><br>
+
+    <button class="submit" type="submit">Login</button>
   </form>
   <p id="error"></p>
   <br><br />
@@ -23,3 +28,5 @@
     </button>
   </Label>
   <script src="../public/js/users.js"></script>
+  <script src="/public/js/helpers.js"></script>
+</main>
