+++ views/dates.handlebars
@@ -1,8 +1,4 @@
-<head>
-  <meta charset="utf-8">
-  <title>{{title}}</title>
-  <link rel="stylesheet" type="text/css" href="../public/css/events.css">
-</head>
+<link rel="stylesheet" type="text/css" href="../public/css/events.css">
 
 <header>
   <h1>{{title}}</h1>
