+++ views/events.handlebars
@@ -1,49 +1,38 @@
-<!DOCTYPE html>
-<html lang="en">
+<link rel="stylesheet" type="text/css" href="../public/css/events.css">
 
-<head>
-  <meta charset="utf-8">
-  <title>{{title}}</title>
-  <link rel="stylesheet" type="text/css" href="../public/css/events.css">
-</head>
+<header>
+  <h1>{{title}}</h1>
 
-<body>
-  <header>
-    <h1>{{title}}</h1>
-
-    <!-- https://www.w3schools.com/howto/howto_css_dropdown_navbar.asp -->
-    <div class="navbar">
-      <a href="/home">Home Page</a>
-      <div class="dropdown">
-        <button class="dropbtn">
-          Events
-          <i class="fa fa-caret-down"></i>
-        </button>
-        <div class="dropdown-content">
-          <a href="/events">All Events</a>
-          <a href="/events/city">Search by Borough</a>
-          <a href="/events/date">Search by Date</a>
-        </div>
+  <!-- https://www.w3schools.com/howto/howto_css_dropdown_navbar.asp -->
+  <div class="navbar">
+    <a href="/home">Home Page</a>
+    <div class="dropdown">
+      <button class="dropbtn">
+        Events
+        <i class="fa fa-caret-down"></i>
+      </button>
+      <div class="dropdown-content">
+        <a href="/events">All Events</a>
+        <a href="/events/city">Search by Borough</a>
+        <a href="/events/date">Search by Date</a>
       </div>
     </div>
-  </header>
+  </div>
+</header>
 
-  <main>
-    <section class="events-list">
-      {{#each events}}
-      <div class="event">
-        <h4>{{this.event_name}}</h4>
-        <ul>
-          <li>Hosts: {{this.event_agency}}</li>
-          <li>Borough: {{this.event_borough}}</li>
-          <li>Date: {{this.start_date_time}}</li>
-          <li>Location: {{this.event_location}}</li>
-          <li>Type: {{this.event_type}}</li>
-        </ul>
-      </div>
-      {{/each}}
-    </section>
-  </main>
-</body>
-
-</html>
+<main>
+  <section class="events-list">
+    {{#each events}}
+    <div class="event">
+      <h4>{{this.event_name}}</h4>
+      <ul>
+        <li>Hosts: {{this.event_agency}}</li>
+        <li>Borough: {{this.event_borough}}</li>
+        <li>Date: {{this.start_date_time}}</li>
+        <li>Location: {{this.event_location}}</li>
+        <li>Type: {{this.event_type}}</li>
+      </ul>
+    </div>
+    {{/each}}
+  </section>
+</main>
