+++ views/home.handlebars
@@ -28,3 +28,5 @@
 
 <script src="/public/js/wifi.js"></script>
 <script src="/public/js/coffee.js"></script>
+<script src="/public/js/helpers.js"></script>
+<script src="/public/js/review.js"></script>
