+++ views/map.handlebars
@@ -53,6 +53,8 @@
     <script src="/public/js/script.js"></script>
     <script src="/public/js/wifi.js"></script>
     <script src="/public/js/coffee.js"></script>
+    <script src="/public/js/helpers.js"></script>
+    <script src="/public/js/review.js"></script>
     {{!--
     <script src="/public/js/map-display.js"></script> --}}
 </body>
